Copy updates from CEL for navgen.

+++ src/plugins/common/recast/recastnavigation/MeshLoaderObj.cpp
@@ -141,7 +141,7 @@ bool rcMeshLoaderObj::load(const char* filename)
     if (!fp)
         return false;
     fseek(fp, 0, SEEK_END);
-    int bufSize = ftell(fp);
+    long bufSize = ftell(fp);
     if (bufSize < 0)
     {
         fclose(fp);
